move get_start_stage from Runner to ConfigExperiment
